pkg/boot/universalpayload: Add ARM64 support
Universal Payload AARCH64 architecture implementation follows same
loading process as X86 architecture, so shares common logic with X86.

To architecture differentation parts, implementation of get Physical
Address Size and Trampoline code reside in architecture specified
source code files.
